LayoutTests/imported/w3c/web-platform-tests/user-timing/test_user_timing_clear_marks-expected.txt

Description

This test validates that the performance.clearMarks() method is working properly. This test creates the following marks to test this method:

"mark1"
"mark2"
After creating each mark, performance.clearMarks() is called three times. First, it is provided with a name of "markUndefined", a non-existent mark, which shouldn't change the state of the Performance Timeline. Next, it is provided with a name of "mark2", after which, this mark should no longer be present in the Performance Timeline. Finally, performance.clearMarks() is called without any name provided. After this call, no marks should be present in the Performance Timeline. The state of the Performance Timeline is tested with the performance.getEntriesByType() and performance.getEntries() methods.

PASS window.performance is defined 
PASS Two marks have been created for this test. 
PASS After a call to window.performance.clearMarks("markUndefined"), where "markUndefined" is a non-existent mark, window.performance.getEntriesByName("mark1") returns an object containing the "mark1" mark. 
PASS After a call to window.performance.clearMarks("markUndefined"), where "markUndefined" is a non-existent mark, window.performance.getEntriesByName("mark2") returns an object containing the "mark2" mark. 
PASS After a call to window.performance.clearMarks("mark1"), window.performance.getEntriesByName("mark1") returns an empty object. 
PASS After a call to window.performance.clearMarks("mark1"), window.performance.getEntriesByName("mark2") returns an object containing the "mark2" mark. 
PASS After a call to window.performance.clearMarks(), window.performance.getEntriesByType("mark") returns an empty object. 
